Laptop price prediction project
The Laptop Price Prediction project leverages a wide range of technologies and machine learning algorithms to create accurate predictions for laptop prices.
The Laptop Price Prediction project leverages a wide range of technologies and machine learning algorithms to create accurate predictions for laptop prices. By using an ensemble of models and a user-friendly interface, this project aims to assist consumers and retailers in making informed decisions related to laptop purchases and pricing strategies. The combination of Python, various machine learning algorithms, and deployment technologies has enabled the development of a powerful and accessible tool for laptop price prediction.